Only scan dirty stacks for minor gc
When stacks are realloc'ed, if the old stack was dirty, the new stack is
marked as dirty. Old stack is also marked as clean since there is no
longer a need to scan the old stack. Since such a dirty old stack will
be on the list of remembered stacks scanned during minor gc, we only
scan stacks if they are dirty. This optimization avoids scanning old
dirty realloc'ed stacks.