remove ocamlbuild -- to conform with other 4.03 switches
Because of the separate-ocamlbuild transition that happens in 4.03,
the opam metadata now expect OCaml version <4.03 to provide
ocamlbuild, and >=4.03 not to. In the current state of the master
branch, I find myself unable to install merlin once the switch is
installed (maybe this means that the constraints I added on
base-ocamlbuild to enforce <4.03 is wrong and should be lifted;
I assumed that experimental switches would follow trunk in removing
ocamlbuild).

The present patch removes ocamlbuild from this switch, so that the
separate ocamlbuild package (0.9.1 currently) can be installed.